Switch to PowerJSON
PowerJSON is a fork of fastJSON, one of the fastest and smallest JSON
serialization libraries. It's relatively new but looks promising. It
seems to have an issue serializing nullable enum types, however, many
serializers so far have had this issue, so perhaps switching to
representing "null" as -1 in the enum is easier for widespread
implementation of the format anyway.